# [zstd_stream] Don't block in reader.Read if a zstd block is available