@delthas 2021-03-30
Benchmarks:
Before *(modified the Decompression test to use io.ReadAll so that we test the same stuff as After)*:
```
goos: windows
goarch: amd64
pkg: github.com/DataDog/zstd
cpu: Intel(R) Core(TM) i7-3820 CPU @ 3.60GHz
BenchmarkStreamCompression
BenchmarkStreamCompression-8     	     248	 143712750 ns/op	  69.38 MB/s
BenchmarkStreamDecompression
BenchmarkStreamDecompression-8   	    2014	  17949798 ns/op	 555.47 MB/s
PASS
```

After:
```
goos: windows
goarch: amd64
pkg: github.com/DataDog/zstd
cpu: Intel(R) Core(TM) i7-3820 CPU @ 3.60GHz
BenchmarkStreamCompression
BenchmarkStreamCompression-8     	     244	 144691207 ns/op	  68.91 MB/s
BenchmarkStreamDecompression
BenchmarkStreamDecompression-8   	    1989	  18085000 ns/op	 551.32 MB/s
PASS
```

No significant change for compression (as expected), ~1% slowdown for decompression. IMO no need to add a flag for that, the performance difference is OK.