@Viq111 2021-06-04
For benchmarking, I think those changes are immaterial so I would not worry about them:
```
ᐅ benchstat
goos: darwin
goarch: amd64
pkg: github.com/DataDog/zstd
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
name                    old time/op    new time/op    delta
CtxCompression-16         46.8µs ± 1%    50.5µs ±13%   ~     (p=0.065 n=9+10)
CtxDecompression-16       2.70µs ± 5%    2.95µs ±15%   ~     (p=0.247 n=10+10)
StreamCompression-16       869ns ±13%     898ns ±12%   ~     (p=0.190 n=10+10)
StreamDecompression-16    2.50µs ± 7%    2.57µs ±14%   ~     (p=0.952 n=10+10)
Compression-16            51.0µs ± 8%    52.1µs ± 8%   ~     (p=0.684 n=10+10)
Decompression-16          3.07µs ±16%    3.11µs ±15%   ~     (p=0.542 n=10+10)

name                    old speed      new speed      delta
CtxCompression-16        151MB/s ± 1%   140MB/s ±12%   ~     (p=0.065 n=9+10)
CtxDecompression-16     2.61GB/s ± 5%  2.42GB/s ±14%   ~     (p=0.247 n=10+10)
StreamCompression-16    8.15GB/s ±12%  7.90GB/s ±11%   ~     (p=0.190 n=10+10)
StreamDecompression-16  2.83GB/s ± 7%  2.76GB/s ±13%   ~     (p=0.971 n=10+10)
Compression-16           139MB/s ± 7%   136MB/s ± 8%   ~     (p=0.684 n=10+10)
Decompression-16        2.31GB/s ±14%  2.28GB/s ±13%   ~     (p=0.529 n=10+10)
```